Добавляет разбиения на модули
